+++ app/Http/Controllers/ActivityLogController.php
@@ -2,12 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Activity;
 use App\Models\Organization;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
-use Spatie\Activitylog\Models\Activity;
 
 class ActivityLogController extends Controller
 {